Fixed GET ERROR_FINISH & REFRESH_PROXY flow
The task type GET had an error in its ERROR_FINISH scenario where it was calling performTask() with the value of "ERROR_FINISH" as the task, not "ERROR_FINISH". 

REFRESH_PROXY in the match task: statement was refreshing the proxy but then doing nothing. This has been changed to repeat the task which called REFRESH_PROXY.